Update to use modern CMake (#17)
This update moves the CMake script for this project over to using modern target based CMake. This main CMakeList.txt is now includable in other projects either via git submodule, CMake Fetch_Content or sub-projecting methods.

The project now has an install and packaging phase, which packages up the contents with config and version information in a platform-independent manner (as this is a header-only library). This is packaged in a relocatable form so that works for Window and Mac where install locations are not fixed.

Code changes by Anthony Peacock - @Twon 

* Add license file

* Update to modern CMake and allow the project to be include, and installed and packaged in relocatable form

* Visual studio compile options fix

* Ensure travis has a recent CMake version installed

* Correct cmake URL for travis

* Expose sanitizers via build script

* Add instructions for building to the README

* Add support for Visual Studio 2017